requirements.txt
streamlit>=1.30.0
arxiv>=2.1.0
semanticscholar>=0.8.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
networkx>=3.2
pyecharts>=2.0.0
streamlit-echarts>=0.4.0
streamlit-agraph>=0.3.0
plotly>=5.18.0
wordcloud>=1.9.0
matplotlib>=3.8.0
openai>=1.10.0
python-dotenv>=1.0.0
